Update Dockerfile to use fabric8/java-centos-openjdk8-jre image
When running REMReM Generate in a Docker container there's little point
in using a Tomcat image (tomcat:8.0-jre8) and running the war file as
the sole application in its Tomcat instance. There are also a few
concrete problems in such a setup:

    - Because the process inside the container with pid 1 won't respond
      to SIGTERM signals, container shutdowns will (by default) take
      ten seconds and the processes will be stopped with SIGKILL,
      leaving no opportunity to clean anything up.
    - Tomcat and the application each have their own logging
      configurations, so introducing a common log format or rotation
      means extra work.

By switching to fabric8/java-centos-openjdk8-jre we run the war file as
a standalone application with the Tomcat that's built into Spring Boot.

This commit introduces a backwards incompatible change, namely that
the directory where the application looks for application.properties
changes.

While updating the Docker documentation we also drop references to the
--expose flag since it's superfluous given the "EXPOSE 8080" instruction
in the dockerfile.